Emacs: fix my-electric-pair-conservative-inhibit bug
'Wrong type argument' exception occurs when typing a '"' at beginning of the buffer
because of the bug fo my-electric-pair-conservative-inhibit.

my-electric-pair-conservative-inhibit determines if a '"' is at the end of a word
by calling (char-syntax (char-before (1- (point))).
char-syntax raises a wrong type argument exception if (point) is less that 2.
We should check the value of (point) before end-of-word test.