tmux.sh: make "home" session's directory to be "$HOME"
If there's no tmux server and you run tmux somewhere which is not the home
directory, tmux will set that directory as "home" session's directory.
This is not probably going to be what you want.

This change makes sure that the "home" session's directory to be exactly the
home directory regardless of the place where you run tmux.