.profile: rename DEFAULT_PATH to INITIAL_PATH
INITIAL_PATH is a better name than DEFAULT_PATH because it reflects
the purpose of the variable more clearly.
It has the very early value of PATH and is used as an initial value of
directory name accumulation.

Also, reset the value of PATH just before accumulate to emphasize
we're rebuilding the value of PATH from the scratch.