Tmux: reflect swap-window behavior change of 3.0
swap-window has changed its behavior from the last major release.
Instead of the window being swapped and the focus following the window that was
originally focussed, the windows are swapped and focus changes to the window it
was swapped with.
If you want to restore the old behavior, you need to give -d option
to swap-window.
For more detail visit: https://github.com/tmux/tmux/issues/1879