taemin.el: introduce taemin-project-compile
Let's say that a makefile is located at the project root directory
and you're editing a file in a sub-directory.
Running taemin-compile in that buffer won't work because it cannot
find a Makefile in that sub-directory.

To build a project, you have to change the buffer to one whose
visited file is at the root directory before compilation or
insert "cd .. &&" in the compile command.
This is annoying.
taemin-project-compile solves this problem by searching the root
directory before compilation.

A project root directory is defined as a directory containing VCS
repo (e.g. git).
If you don't use VCS, just put an empty .projectile file in the directory.