untitled-note.el: turn off untitled-mode after write-file
If you write an untitled note to the file which has "txt" as the file extension,
like foo.txt, the untitled-note mode won't disappear even after writing to a
file. Because the major-mode is the same as the old one. This problem has
introduced by commit 8e026c8.

This patch solves this issue by turning off untitled-note mode forcible after
writing to a file.