fzf: set up keybindings and auto completion for MacPorts fzf
The keybindings and auto completion are automatically set up if you build
and install fzf manually from the source. But that's not the case if you
install fzf via a package manager like Homebrew or MacPorts.

This patch teaches Mac/setup to setup the keybindings and auto-completion
even if you install fzf via a package manager.