Tmux: Open a new pane with current directory
Currently, every time a new Tmux pane is opened, it defaults to the home directory.
This behavior can be frustrating for users who frequently switch between directories.

To address this issue, we need to teach Tmux to open a new pane in the same directory as the current pane.
By doing so, users will be able to save time and avoid unnecessary cd.

Please refer to the following link for more information:
https://gist.github.com/william8th/faf23d311fc842be698a1d80737d9631