Tmux: Resolve window shifting bug
The keybindings <prefix> M-Left and M-Right are used for window shifting,
which allows users to shift the current window left and right respectively.
However, this functionality has a bug that affects windows at the very left
or right corner.

For instance, consider the case of four windows: a, b, c, and x, where x is
the current window.

a b c [x]

When we left-shift window x, it moves through the left side of the screen in
a cyclic manner, as follows:

a b [x] c
a [x] b c
[x] a b c

But when we try to shift x one more time, the bug becomes apparent:
c moves to the left corner.

c a b [x]

Obviously, that's not what we want. What we expect is:

a b c [x]

To fix this issue, we need to add appropriate considerations when going
through the boundary.  Currently, the window shifting is implemented by
swap-window which doesn't account for boundary conditions.  This patch
teaches Tmux to select the appropriate command between swap-window and
move-window depending on the current window's location.