Emacs: defer glsl-mode loading
use-package gives a glsl-mode not found errer message on every startup if
glsl-mode is missing. That's because there is no keyword like :commands,
:bind, :mode, :interpreter or :hook (all of which imply :defer t) in the
glsl-mode settings. See https://github.com/jwiegley/use-package#modes-and-interpreters

This patch teaches Emacs to defer loading of glsl-mode until the first
visit to a glsl file to prevent the error message on startup.