Shell: Unset empty MANPATH
When running Zsh on OSX, the system-wide startup files define MANPATH as an
empty string. This can cause issues when external tools modify the MANPATH
environment variable.  For example, `brew shellenv` gives me an output
like below:

  export HOMEBREW_PREFIX="/usr/local";
  export HOMEBREW_CELLAR="/usr/local/Cellar";
  export HOMEBREW_REPOSITORY="/usr/local/Homebrew";
  export PATH="/usr/local/bin:/usr/local/sbin${PATH+:$PATH}";
  export MANPATH="/usr/local/share/man${MANPATH+:$MANPATH}:";
  export INFOPATH="/usr/local/share/info:${INFOPATH:-}";

MANPATH is set but empty and you eval that output, it is going to be
"/usr/local/share/man::".

To avoid these unexpected results, this patch teaches .zprofile to check if
MANPATH is defined but empty and unset it if necessary. While this change
may not appear to have a significant impact, it ensures a consistent and
predictable environment for users.