Zsh: Match Ctrl-U behavior with Bash
In Zsh, the default behavior of Ctrl-U is to delete the entire line. But
deleting the entire line is not as commonly needed as deleting line forward
or backward. More importantly, it differs from Bash's behavior.

Let's bind backward-kill-line to Ctrl-U to match the behavior with Bash. We
can still delete the entire line by Ctrl-A Ctrl-K, or simply by Ctrl-C if
you just want to discard the line.