Blank: Introduce custom kill commands to restore placeholder
When you press DEL key on a blank, it just deletes a single character without
restoring the blank placeholder. This behavior can make it difficult for user to
identify the extent of the blank. The same issue applies to other deletion
commands which are bound to C-d, M-d, and M-DEL keys.

To resolve this issue, this patch introduces customized kill commands and binds
them to above keys. The customized commands not only kill characters but also
restore the placeholders.