Aspell: Include personal dictionaries
This patch adds Aspell personal dictionaries to teach Aspell to recognize words
that are not present in the official dictionary.

The personal dictionary will now include words such as "dotfile", "dotfiles",
and "utf".

For futher information on the format and usage of personal dictionaries, please
refer to the following link:
http://aspell.net/man-html/Format-of-the-Personal-and-Replacement-Dictionaries.html