Sublimey: Refine paragraph commands
By default, Emacs considers the beginning of a blank line that precedes
paragraph to be the start of a paragraph. However, SublimeText doesn't. It
considers the beginning of the first line of paragraph to be start of the
paragraph.

This patch introduces new paragraph commands that use the SublimeText's
approach for better emulation.