Sublimey: Leap over empty lines when going to next word
The function sublimey-forward-word goes to the next line if cursor is at the end
of the line or the line is empty. It is an acceptable behavior and the same as
Sublime Text. But this behavior becomes frustrating when the cursor meets
multiple empty lines. Because the user has to press the key many times to go to
the next non-empty line. The same thing goes to the functions
sublimey-backward-word, sublimey-kill-word, and sublimey-backward-kill-word.

This patch teaches these functions to leap over or kill the sequence of empty
lines at once if sublimey-word-leap-empty-lines is t.