Emacs: Facilitate window navigation with Hydra
On MacBook, I remapped the left arrow key to right control key. So windmove
default keybindings(shift-arrow) is not effective for me. To work around this,
This patch makes use of hydra.

With prefix C-x o, we can go to the window we want by repeating keys c, t, h or
n, like C-x o c c n to move up up and right. The keys c, t, h, and n sit on the
home row of Dvorak keyboard. So no need to move our hand to press them. This
approach may be a little bit cumbersome than using shift-arrow keys, but it is
acceptable enough.