Blank: Robust word checking
blank-hide-random-words checks if it is possible to blank a word by
using (looking-at-p "\\w\\w"). However, this method returns t when encountering
a digit followed by word characters, such as "1word", which can lead to failure
in filtering out words that cannot be blanked.

To resolve this issue, this patch introduces an additional argument that
represents the position of word's end to the predicate. With this argument, it
becomes easier for the caller to determine if the word is good or not. The
length check can be achieved by simply subtracting two arguments.