Emacs: Replace shell with ansi-term
Emacs shell gets mad easily with ncurse. ansi-term handles it better.

This patch introduces three interactive functions called taemin-terminal,
taemin-terminal-other-window, and taemin-terminal-other-window-reuse.

taemin-terminal create a terminal buffer and switches to it.

taemin-terminal-other-window does the same, but open it in other window.
This function is bound to <S-f6>.

taemin-terminal-other-window-reuse enables to reuse the already opened
terminal. It's bound to <f6>.

A Terminal will quit automatically when the shell exits. If you are visiting
other buffers, then the terminal will remain. You can later quit the window
by pressing the 'q' key.

Implementation Note:

It's worth noting that directly replacing switch-to-buffer in the ansi-term
code with switch-to-buffer-other-window doesn't work flawlessly, especially
with Zsh. This can lead to an issue that a highlighted percent sign appears
above the prompt if switch-to-buffer-other-window splits the window
side-by-side, as shown in the image below:

  https://user-images.githubusercontent.com/60341709/121371372-158cc800-c970-11eb-827a-2bc83aa22730.png

We can solve this issue by deferring the terminal initialization until after
the buffer switching.