Emacs: Jump to anywhere with few candidates
avy-goto-subword-1 gives fewer candidates than avy-goto-char but it is limited
to moving only to the beginning of a sub-word. On the other hand, avy-goto-char
can go to anywhere but may yield too many candidates.

To strike a better balance between these functions, we switch to
avy-goto-char-timer. This function has the advantage of both of them. It allows
navigation to anywhere giving us fewer candidates by receiving arbitrary
number of characters to match.

Let's use avy-got-char-timer instead of avy-goto-subword-1.