Emacs: Simply key sequences to reduce modifiers
Some of frequently-used key sequences, such as C-x C-f or C-x C-s are cumbersome
because of their modifiers. It's hard to remember and type in. Maybe we can't
avoid the first modifier but can remove modifiers from the following key
sequence. ex) C-x C-f => C-x f

set-fill-column is bound to C-x f, but I never use it. And delete-window is
bound to C-x w and C-x 0, but I always go to C-x 0. So it's safe to bind to C-x
f and C-x w.