Emacs: Facilitate window resizing with Hydra
Without numeric prefix, enlarge-window enlarges the current window by 1 line
height. So we have to press C-x ^ multiple times to resize the window to the
desired size, which is annoying.

This patch introduces a hydra to make it possible to resize a window by
repeating a single keystroke. For example, to enlarge window, you press C-x and
then press ] as many times as needed to reach the desired size. When the window
is large enough, you press RET to finalize the resizing. The same approach
applies to shrinking the window.