Git: Allow to fixup a past commit
Currently, "git fix" and "git amend" are defined as aliases and they can
only handle the most recent commit. You need to rebase yourself if you want
to deal with a past commit.

To address this limitation, This patch introduces a new Git custom command
called "git-fix". It can fixup, amend, or reword not only the most recent
commit but also a commit in the past.