Git: Allow to fixup a past commit
Currently, "git fix" and "git amend" are defined as aliases and they only
handle the most recent commit. So we can't use them if the commit we want to
deal with is in the past.

To address this limitation, This patch introduces a new Git custom command
called "git-fix". It can fixup, amend, or reword not only the most recent
commit but also a commit in the past.

Like "git commit", "git-fix" also supports --all option (-a in short) to
stage modified or deleted files before fixing up. However you have to give
the option before the target commit. So

  git fix -a abcde # okay
  git fix abcde -a # This will ignore -a