Install 'z' man page
bootstrap.sh downloads and installs 'z' directory jump utility in the home
directory, but it doesn't install the man page.

This patch teaches bootstrap.sh to install the man page in the directory
~/.local/share/man/man1 which is XDG standard.

bootstrap.sh creates ~/.local/bin and .profile adds that directory to PATH.
man is smart enought to infer the man page directories from PATH variable.
So we don't need to touch the MANPATH variable.