Tmux: Open new window adjacent to the current
Commit 50e7056c("Tmux: Open a new pane with current directory", 2023-04-17) had
the new window to inherit the current path to make it easy for user to do a job
in another window. But the new window pops up at the end of the window tab
without a designated name. So the user has to manually rearrange the windows and
rename it for better organization.

This patch teaches Tmux to position the new window immediately after the current
one and also inherit its window name to make it easy to manage windows.