Emacs: Preserve forward-paragraph behavior in sgml-mode
sgml-mode redefines paragraph-start and paragraph-separate to make
forward-paragraph and backward-paragraph go to the beginning or end of a tag
such as, <p> or <h1> which generates a text block separated from adjacent blocks
by blank. This looks like a good idea at first glance, but it is not. This
behavior breaks user's expectations. forward-paragraph and backward-paragraph
should only consider the HTML code itself instead of contents it will generate.

This patch prevents sgml-mode from redefining paragraph by adding a hook which
deletes the local variables paragraph-start and paragraph-separate. With this
update, forward-paragraph and backward-paragraph behave as expected.