Emacs: Open terminal in a separate window
When opening a terminal, it's often convenient to have it appears in a
separate window. This allows you to continue viewing the current buffer
while working in the terminal. But, taemin-terminal just switches the buffer
of the current window instead of opening a new one.

This patch introduces an additional function, taemin-terminal-other-window,
designed to open a terminal in a separate window. This function is the
default behavior when you press the F6 key.

Implementation Note:

It's worth noting that directly replacing switch-to-buffer in the ansi-term
code with switch-to-buffer-other-window doesn't work flawlessly, especially
with Zsh. This can lead to an issue that a highlighted percent sign appears
above the prompt if switch-to-buffer-other-window splits the window
side-by-side, as shown in the image below:

  https://user-images.githubusercontent.com/60341709/121371372-158cc800-c970-11eb-827a-2bc83aa22730.png

We can solve this issue by deferring the terminal initialization until after
the buffer switching.