Emacs: Use Ivy-like pattern matching when searching
Ivy provides a convenient and relatively fast pattern match method called
`ivy--regex-plus' which treats whitespace in the pattern as ".*". We can mimic
this behavior when incremental searching by defining `search-whitespace-regexp'
as ".*?".

`occur' doesn't respect `search-whitespace-regexp' so this commit defines a
customized version of it called `taemin-occur'.

With this change, we no longer need Swiper. Let's remove it.