Create alias to tmux only if tmux is found
bootstrap.sh always copies tmux.sh to the home directory even if tmux is not
found. This makes it hard to determine if tmux is available by `command -v tmux`
because it always says "alias tmux='_tmux'".

This commit teaches bootstrap.sh to only copy tmux.sh when tmux is available.