Emacs: Make it easy to copy buffer file name
You can check the file name of the visiting buffer by evaluating
buffer-file-name, but it requires a bunch of typings and you have to switch
to *Message* buffer if you want to copy it.

This commit introduces a function to copy the file name of the buffer. Press
M-x and typing "buffer filename" requires less effort than evaluating
buffer-file-name.

Press C-u before running function if you just want to see the file name, but
no copy it.